Refactor pfcwd orch code.
