+++ orchagent/pfcwdorch.h
@@ -14,6 +14,7 @@ extern "C" {
 }
 
 #define PFC_WD_FLEX_COUNTER_GROUP       "PFC_WD"
+#define PFC_WD_TC_MAX                   8
 
 const string pfc_wd_flex_counter_group = PFC_WD_FLEX_COUNTER_GROUP;
 
@@ -31,12 +32,11 @@ static const map<string, sai_packet_action_t> packet_action_map = {
     {"alert", SAI_PACKET_ACTION_FORWARD}
 };
 
-template <typename DropHandler, typename ForwardHandler>
-class PfcWdOrch: public Orch
+class PfcWdBaseOrch: public Orch
 {
 public:
-    PfcWdOrch(DBConnector *db, vector<string> &tableNames);
-    virtual ~PfcWdOrch(void);
+    PfcWdBaseOrch(DBConnector *db, vector<string> &tableNames);
+    virtual ~PfcWdBaseOrch(void);
 
     virtual void doTask(Consumer& consumer);
     virtual bool startWdOnPort(const Port& port,
@@ -54,7 +54,7 @@ public:
     }
 
     static PfcWdAction deserializeAction(const string& key);
-    static string serializeAction(const PfcWdAction &action); 
+    static string serializeAction(const PfcWdAction &action);
 
     virtual task_process_status createEntry(const string& key, const vector<FieldValueTuple>& data);
     task_process_status deleteEntry(const string& name);
@@ -63,6 +63,19 @@ public:
 
 protected:
     virtual bool startWdActionOnQueue(const string &event, sai_object_id_t queueId, const string &info="") = 0;
+
+    // Get lossless TCs for a port based on PFC mask
+    bool getLosslessTcsForPort(const Port& port, set<uint8_t>& losslessTc);
+
+    // Report PFC storm event
+    void report_pfc_storm(sai_object_id_t queueId, sai_object_id_t portId,
+                         uint8_t queueIndex, const string& portAlias,
+                         const string& info = "");
+
+    // Report PFC storm restored event
+    void report_pfc_restored(sai_object_id_t queueId, sai_object_id_t portId,
+                            uint8_t queueIndex, const string& portAlias);
+
     string m_platform = "";
     shared_ptr<FlexCounterTaggedCachedManager<sai_object_type_t>> m_pfcwdFlexCounterManager;
 
@@ -72,81 +85,6 @@ private:
     shared_ptr<Table> m_countersTable = nullptr;
     PfcWdAction PfcDlrPacketAction = PfcWdAction::PFC_WD_ACTION_UNKNOWN;
     std::set<std::string> m_pfcwd_ports;
-};
-
-template <typename DropHandler, typename ForwardHandler>
-class PfcWdSwOrch: public PfcWdOrch<DropHandler, ForwardHandler>
-{
-public:
-    PfcWdSwOrch(
-            DBConnector *db,
-            vector<string> &tableNames,
-            const vector<sai_port_stat_t> &portStatIds,
-            const vector<sai_queue_stat_t> &queueStatIds,
-            const vector<sai_queue_attr_t> &queueAttrIds,
-            int pollInterval);
-    virtual ~PfcWdSwOrch(void);
-
-    void doTask(Consumer& consumer) override;
-    virtual bool startWdOnPort(const Port& port,
-            uint32_t detectionTime, uint32_t restorationTime, PfcWdAction action, string pfcStatHistory);
-    virtual bool stopWdOnPort(const Port& port);
-
-    task_process_status createEntry(const string& key, const vector<FieldValueTuple>& data) override;
-    virtual void doTask(SelectableTimer &timer);
-    //XXX Add port/queue state change event handlers
-
-    bool bake() override;
-    void doTask() override;
-
-protected:
-    bool startWdActionOnQueue(const string &event, sai_object_id_t queueId, const string &info="") override;
-
-private:
-    struct PfcWdQueueEntry
-    {
-        PfcWdQueueEntry(
-                PfcWdAction action,
-                sai_object_id_t port,
-                uint8_t idx,
-                string alias);
-
-        PfcWdAction action = PfcWdAction::PFC_WD_ACTION_UNKNOWN;
-        sai_object_id_t portId = SAI_NULL_OBJECT_ID;
-        uint8_t index = 0;
-        string portAlias;
-        shared_ptr<PfcWdActionHandler> handler = { nullptr };
-    };
-
-    template <typename T>
-    static unordered_set<string> counterIdsToStr(const vector<T> ids, string (*convert)(T));
-    bool registerInWdDb(const Port& port,
-            uint32_t detectionTime, uint32_t restorationTime, PfcWdAction action, string pfcStatHistory);
-    void unregisterFromWdDb(const Port& port);
-    void doTask(swss::NotificationConsumer &wdNotification);
-
-    unordered_set<string> filterPfcCounters(const unordered_set<string> &counters, set<uint8_t>& losslessTc);
-    string getFlexCounterTableKey(string s);
-
-    void disableBigRedSwitchMode();
-    void enableBigRedSwitchMode();
-    void setBigRedSwitchMode(string value);
-
-    void report_pfc_storm(sai_object_id_t id, const PfcWdQueueEntry *, const string&);
-
-    map<sai_object_id_t, PfcWdQueueEntry> m_entryMap;
-    map<sai_object_id_t, PfcWdQueueEntry> m_brsEntryMap;
-
-    const vector<sai_port_stat_t> c_portStatIds;
-    const vector<sai_queue_stat_t> c_queueStatIds;
-    const vector<sai_queue_attr_t> c_queueAttrIds;
-
-    bool m_bigRedSwitchFlag = false;
-    int m_pollInterval;
-
-    shared_ptr<DBConnector> m_applDb = nullptr;
-    // Track queues in storm
-    shared_ptr<Table> m_applTable = nullptr;
 };
 
 #endif
