+++ tests/mock_tests/mock_orchagent_main.h
@@ -19,6 +19,7 @@
 #include "qosorch.h"
 #define protected public
 #include "pfcwdorch.h"
+#include "pfcwdsworch.h"
 #undef protected
 #undef private
 #include "vrforch.h"
